bug fix
timeline infinit reload bug fixed
Delete account bug fixed

+++ lib/features/authentication/repos/auth_repo.dart
@@ -41,6 +41,10 @@ class AuthenticationRepository {
   }
 
   Future<void> deleteAccount() async {
+    if (user == null) {
+      signOut();
+      return;
+    }
     await user!.delete();
     await signOut();
   }
